refactor: extract hdf5 io helpers
