+++ dfode_kit/__init__.py
@@ -32,8 +32,8 @@ _ATTRIBUTE_MODULES = {
     "inverse_BCT_torch": ("dfode_kit.utils", "inverse_BCT_torch"),
     "gather_species_arrays": ("dfode_kit.df_interface.sample_case", "gather_species_arrays"),
     "df_to_h5": ("dfode_kit.df_interface.sample_case", "df_to_h5"),
-    "touch_h5": ("dfode_kit.data_operations.h5_kit", "touch_h5"),
-    "get_TPY_from_h5": ("dfode_kit.data_operations.h5_kit", "get_TPY_from_h5"),
+    "touch_h5": ("dfode_kit.data.io_hdf5", "touch_h5"),
+    "get_TPY_from_h5": ("dfode_kit.data.io_hdf5", "get_TPY_from_h5"),
     "advance_reactor": ("dfode_kit.data_operations.h5_kit", "advance_reactor"),
     "load_model": ("dfode_kit.data_operations.h5_kit", "load_model"),
     "predict_Y": ("dfode_kit.data_operations.h5_kit", "predict_Y"),
