+++ dfode_kit/cli_tools/commands/sample.py
@@ -30,7 +30,7 @@ def add_command_parser(subparsers):
 
 
 def handle_command(args):
-    from dfode_kit.data_operations.h5_kit import touch_h5
+    from dfode_kit.data.io_hdf5 import touch_h5
     from dfode_kit.df_interface.sample_case import df_to_h5
 
     print('Handling sample command')
